% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VerifyEpName.R
\name{.VerifyEpName}
\alias{.VerifyEpName}
\title{Should be inputted as one row per person for CR data.
 Not needed if splitRule = csh_cr for endpoint 'CR'}
\usage{
.VerifyEpName(epName, ...)
}
\description{
Should be inputted as one row per person for CR data.
 Not needed if splitRule = csh_cr for endpoint 'CR'
}
