% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VerifySplitRule.R
\name{.VerifySplitRule,NULL-method}
\alias{.VerifySplitRule,NULL-method}
\title{If NULL and Phase = 1 and 'criticalValue' = 'mean', then takes value 'mean'.
  If NULL and Phase = 1 and 'criticalValue' = 'prob' or 'mean.prob.combo', then takes value 'logrank'.
  If NULL and Phase = 2 and endPoint = 'CR', takes value 'gray_cr'. # default over csh_cr
  If NULL and Phase = 2 and endPoint = 'RE', takes value 'gray_re'.}
\usage{
\S4method{.VerifySplitRule}{`NULL`}(splitRule, ..., Phase, endPoint, criticalValue)
}
\description{
If NULL and Phase = 1 and 'criticalValue' = 'mean', then takes value 'mean'.
  If NULL and Phase = 1 and 'criticalValue' = 'prob' or 'mean.prob.combo', then takes value 'logrank'.
  If NULL and Phase = 2 and endPoint = 'CR', takes value 'gray_cr'. # default over csh_cr
  If NULL and Phase = 2 and endPoint = 'RE', takes value 'gray_re'.
}
