% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VerifySplitRule.R
\name{.VerifySplitRule}
\alias{.VerifySplitRule}
\title{Must be one of \{'logrank_surv', 'mean_surv', 'gray_cr', 'csh_cr', 'gray_re'\}
  indicating the test used to determine an optimal split.
  If NULL and Phase = 'Survival' and 'criticalValue' = 'mean', then takes value 'mean'.
  If NULL and Phase = 'Survival' and 'criticalValue' = 'prob' or 'mean.prob.combo', then takes value 'logrank'.
  If NULL and Phase = 'CR' and endPoint = 'CR', takes value 'graycr'. # default over cshcr
  If NULL and Phase = 'RE' and endPoint = 'RE', takes value 'grayre'.
  ## survival endpoint ##
  logrank = 1
  mean = 2
  ## other endpoints (CR, RE) ##
  cr_gray = 3
  cr_csh = 4
  re_qlr = 5
  if splitRule <=2 then Phase 1 survival
  if splitRule 3-4 then Phase 2 endpoint (CR)
  if splitRule >4 then Phase 2 endpoint (RE)}
\usage{
.VerifySplitRule(splitRule, ...)
}
\description{
Must be one of \{'logrank_surv', 'mean_surv', 'gray_cr', 'csh_cr', 'gray_re'\}
  indicating the test used to determine an optimal split.
  If NULL and Phase = 'Survival' and 'criticalValue' = 'mean', then takes value 'mean'.
  If NULL and Phase = 'Survival' and 'criticalValue' = 'prob' or 'mean.prob.combo', then takes value 'logrank'.
  If NULL and Phase = 'CR' and endPoint = 'CR', takes value 'graycr'. # default over cshcr
  If NULL and Phase = 'RE' and endPoint = 'RE', takes value 'grayre'.
  ## survival endpoint ##
  logrank = 1
  mean = 2
  ## other endpoints (CR, RE) ##
  cr_gray = 3
  cr_csh = 4
  re_qlr = 5
  if splitRule <=2 then Phase 1 survival
  if splitRule 3-4 then Phase 2 endpoint (CR)
  if splitRule >4 then Phase 2 endpoint (RE)
}
