% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ITRSurv.R, R/itrSurv.R
\name{stage,ANY-method}
\alias{stage,ANY-method}
\alias{itrSurv-internal-api}
\title{Retrieve Stage Results as a List}
\usage{
\S4method{stage}{ANY}(object, ...)
}
\description{
Retrieve Stage Results as a List

Hidden methods
}
\keyword{internal}
